refactor(xcode): Moved PackageResolvedTransformer.
